requirements.txt
boto3
Django>=5.0.1
djangorestframework>=3.14.0
django-filter>=23.5
django-debug-toolbar>=4.2.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
django-allauth>=0.62.1
Werkzeug>=2.3.7
zappa>=0.59.0
python-dotenv>=1.0.1
jinja2>=3.1.3
django-chroniker>=1.0.24
psutil>=6.1.0
django-storages[s3]>=1.14.4
fpdf2>=2.7.6
azure-identity==1.25.1
azure-mgmt-authorization==4.0.0
azure-mgmt-costmanagement==4.0.0
azure-mgmt-managementgroups==1.0.0
azure-mgmt-policyinsights==1.0.0
azure-mgmt-resource==24.0.0
azure-mgmt-security==7.0.0
azure-mgmt-subscription==3.1.1